pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "natural-language-workflow-platform"
version = "0.1.0"
description = "A natural language workflow automation platform using Pydantic AI"
authors = [
    {name = "Development Team", email = "dev@example.com"},
]
readme = "README.md"
requires-python = ">=3.11,<4.0"

dependencies = [
    "pydantic-ai[ag-ui]>=0.4.0",
    "openai>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "asyncpg>=0.29.0",
    "redis>=5.0.0",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "aiofiles>=23.2.1",
    "pydantic-settings>=2.0.0",
    "temporalio>=1.6.0",
    "gunicorn>=21.2.0",
    "pydantic>=2.5.0",
    "sqlalchemy>=2.0.0",
    "jinja2>=3.1.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[tool.black]
line-length = 88
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.ruff]
line-length = 88
target-version = "py311"